Comp: Version Control - Cameron James Merged the Version Branch with Trunk.  This has the version control changes included.

+++ 0.3/Apromore-Core/apromore-service/src/main/java/org/apromore/canoniser/adapters/pnml2canonical/TranslateEdgeAnnotations.java
@@ -1,62 +1,60 @@
 package org.apromore.canoniser.adapters.pnml2canonical;
 
-import java.math.BigInteger;
-import java.util.List;
-
-
 import org.apromore.anf.GraphicsType;
 import org.apromore.anf.LineType;
 import org.apromore.anf.PositionType;
 
+import java.util.List;
+
 
 public class TranslateEdgeAnnotations {
-	DataHandler data;
+    DataHandler data;
 
-	public void setValues(DataHandler data) {
-		this.data = data;
-	}
+    public void setValues(DataHandler data) {
+        this.data = data;
+    }
 
-	public void addEdgeAnnotations(Object obj) {
-		GraphicsType graphT = new GraphicsType();
-		LineType line = new LineType();
-		String cpfId = null;
+    public void addEdgeAnnotations(Object obj) {
+        GraphicsType graphT = new GraphicsType();
+        LineType line = new LineType();
+        String cpfId = null;
 
-		org.apromore.pnml.ArcType element = (org.apromore.pnml.ArcType) obj;
+        org.apromore.pnml.ArcType element = (org.apromore.pnml.ArcType) obj;
 
-		cpfId = data.get_id_map_value(element.getId());
+        cpfId = data.get_id_map_value(element.getId());
 
-		if (element.getGraphics() != null) {
+        if (element.getGraphics() != null) {
 
-			if (element.getGraphics().getLine() != null) {
-				line.setColor(element.getGraphics().getLine().getColor());
-				line.setShape(element.getGraphics().getLine().getShape());
-				line.setStyle(element.getGraphics().getLine().getStyle());
-				line.setWidth(element.getGraphics().getLine().getWidth());
-				graphT.setLine(line);
-			}
+            if (element.getGraphics().getLine() != null) {
+                line.setColor(element.getGraphics().getLine().getColor());
+                line.setShape(element.getGraphics().getLine().getShape());
+                line.setStyle(element.getGraphics().getLine().getStyle());
+                line.setWidth(element.getGraphics().getLine().getWidth());
+                graphT.setLine(line);
+            }
 
-			if (element.getGraphics().getPosition() != null) {
-				List<org.apromore.pnml.PositionType> pos1 = element
-						.getGraphics().getPosition();
-				for (Object obj1 : pos1) {
-					org.apromore.pnml.PositionType position = (org.apromore.pnml.PositionType) obj1;
-					PositionType pos = new PositionType();
-					if (position.getX() != null) {
-						pos.setX(position.getX());
+            if (element.getGraphics().getPosition() != null) {
+                List<org.apromore.pnml.PositionType> pos1 = element
+                        .getGraphics().getPosition();
+                for (Object obj1 : pos1) {
+                    org.apromore.pnml.PositionType position = (org.apromore.pnml.PositionType) obj1;
+                    PositionType pos = new PositionType();
+                    if (position.getX() != null) {
+                        pos.setX(position.getX());
 
-					}
-					if (position.getY() != null) {
-						pos.setY(position.getY());
-					}
+                    }
+                    if (position.getY() != null) {
+                        pos.setY(position.getY());
+                    }
 
-					graphT.getPosition().add(pos);
-				}
+                    graphT.getPosition().add(pos);
+                }
 
-			}
+            }
 
-			graphT.setCpfId(BigInteger.valueOf(Long.valueOf(cpfId)));
-			data.getAnnotations().getAnnotation().add(graphT);
-		}
-	}
+            graphT.setCpfId(cpfId);
+            data.getAnnotations().getAnnotation().add(graphT);
+        }
+    }
 
 }
